pyproject.toml
[project]
name = "er-nl2sql"
version = "0.1.0"
description = "Convert questions written in Natural Language to SQL queries"
authors = [
    {name = "ksj1368",email = "ksj006151@gmail.com"}
]
readme = "README.md"
packages = [{ include = "scripts" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "openai==1.25.0",
    "perplexipy>=1.2.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "pymysql>=1.1.0,<2.0.0",
    "python-dotenv>=1.1.0",
    "requests (>=2.32.3,<3.0.0)",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
